int cast on byte doesnt work as it would be expected, so implemented own int2byte cast, improved some parts a lot, but still doesnt work completely

+++ src/test/java/des/BitBlockTest.java
@@ -8,7 +8,7 @@ public class BitBlockTest {
     void getNthBitTest() {
         byte[] bytes = new byte[8];
         bytes[3] = (byte) (0b10000000);
-        byte result = BitBlock.getNthBit(bytes, 25);
+        int result = BitBlock.getNthBit(bytes, 25);
         Assertions.assertEquals(1, result);
     }
 }
